requirements.txt
requests
python-dotenv
pyttsx3
playwright